# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Projeto Integrador Senac",
    "description": "Esta API tem como objetivo facilitar a criação de avisos para os discentes do Mediotech Senac, assim como o gerenciamento de suas notas por parte do professor.",
    "version": "1.0.0",
    "contact": {
      "email": "rogercaua@outlook.com.br"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Servidor local de desenvolvimento"
    },
    {
      "url": "https://backendpi-qkmq.onrender.com",
      "description": "Servidor Publico"
    }
  ],
  "paths": {
...
    "/disciplina/create": {
      "post": {
        "tags": ["Disciplinas"],
        "summary": "Criar uma nova disciplina",
        "operationId": "createDisciplina",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateDisciplina"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Disciplina criada com sucesso"
          },
          "404": {
            "description": "Professor não encontrado"
          },
          "500": {
            "description": "Erro ao criar disciplina"
          }
        }
      }
    },
    "/disciplina": {
      "get": {
        "tags": ["Disciplinas"],
        "summary": "Listar todas as disciplinas",
        "operationId": "getAllDisciplinas",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de disciplinas",
            "schema": {
              "type": "array"
            }
          },
          "500": {
            "description": "Erro ao buscar disciplinas"
          }
        }
      }
    },
    "/disciplina/delete": {
      "delete": {
        "tags": ["Disciplinas"],
        "summary": "Deletar uma disciplina",
        "operationId": "deleteDisciplina",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Disciplina deletada com sucesso"
          },
          "404": {
            "description": "Disciplina não encontrada"
          },
          "500": {
            "description": "Erro ao deletar disciplina"
          }
        }
      }
    },
    "/disciplina/delete/all": {
      "delete": {
        "tags": ["Disciplinas"],
        "summary": "Deletar todas as disciplinas",
        "operationId": "deleteAllDisciplinas",
        "responses": {
          "200": {
            "description": "Todas as disciplinas foram deletadas com sucesso"
          },
          "500": {
            "description": "Erro ao deletar disciplinas"
          }
        }
      }
    },
...
  },
...
}